% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aprofFun.R
\name{aprofFun}
\alias{aprofFun}
\title{Light wrapper around the aprof function to profile one function}
\usage{
aprofFun(expr, FUN)
}
\arguments{
\item{expr}{Expression to profile}

\item{FUN}{The name of the function to profile, as a string (e.g. "f")}
}
\value{
An aprof object
}
\description{
Light wrapper around the aprof function to profile one function
}
\examples{
# The function to profile
f = function(n, sim) {
  # Simulate some data
  x = rnorm(n, mean = 0, sd = 2)
  t = data.frame(x)
  for (i in 1:sim) {
   y = x + rnorm(x, mean = 0, sd = 0.5)
    t = cbind(t, y)
  }
  return(t)
}

# Another function calling f
g = function(x) {
  print(x + x)
  output = f(100, 100)
  return(output)
}

# Profiling:
## f is run
prof = aprofFun(f(100, 100), "f")
viewProfHtml(prof)
## g is run, profiling either "g" or "f"
prof = aprofFun(g(10), "g")
viewProfHtml(prof)
prof = aprofFun(g(10), "f")
viewProfHtml(prof)

}

