% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aprofWrapper.R
\name{aprofWrapper}
\alias{aprofWrapper}
\title{Light wrapper around the aprof function}
\usage{
aprofWrapper(expr, sourceFile)
}
\arguments{
\item{expr}{Expression to profile}

\item{sourceFile}{The file with the source code to be profiled}
}
\value{
An aprof object
}
\description{
Light wrapper around the aprof function
}
\examples{
# The function to profile
f = function(n, sim) {
  # Simulate some data
  x = rnorm(n, mean = 0, sd = 2)
  t = data.frame(x)
  for (i in 1:sim) {
   y = x + rnorm(x, mean = 0, sd = 0.5)
    t = cbind(t, y)
  }
  return(t)
}

# Dump the function to a tmp file
mySourceFile = tempfile()
dump("f", file = mySourceFile)

# Profile
prof = aprofWrapper(f(100, 100), mySourceFile)
# Visualise
viewProfHtml(prof)

}

