% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertAprofToHtml.R
\name{convertAprofToHtml}
\alias{convertAprofToHtml}
\title{Convert an aprof object into an html output}
\usage{
convertAprofToHtml(prof, htmlFile)
}
\arguments{
\item{prof}{aprof object}

\item{htmlFile}{string, path to the output file}
}
\description{
Convert an aprof object into an html output
}

