% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePngLine.R
\name{makePngLine}
\alias{makePngLine}
\title{Make a small embedded png line from a numerical value}
\usage{
makePngLine(value, xmax, alt = "")
}
\arguments{
\item{value}{numerical value to draw}

\item{xmax}{maximum possible value (for scale)}

\item{alt}{alternative text for the image html code}
}
\value{
A string with the embedded image, ready to be inserted in an html
  document
}
\description{
Make a small embedded png line from a numerical value
}

