% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewProfHtml.R
\name{viewProfHtml}
\alias{viewProfHtml}
\title{Main function to view an aprof object as an html page}
\usage{
viewProfHtml(prof)
}
\arguments{
\item{prof}{aprof object}
}
\description{
Main function to view an aprof object as an html page
}

